seperated models from app
to increase the readability and seperate the logic
