+++ models.py
@@ -1,0 +1,19 @@
+# models.py, trying sommething..lets see
+from flask_sqlalchemy import SQLAlchemy
+from flask_login import UserMixin
+import bcrypt
+
+db = SQLAlchemy()
+
+class Users(db.Model, UserMixin):
+    id = db.Column(db.Integer, primary_key=True)
+    username = db.Column(db.String(80), unique=True, nullable=False)
+    email = db.Column(db.String(120), unique=True, nullable=False)
+    password = db.Column(db.String(200), nullable=False)
+    role = db.Column(db.String(20), nullable=False, default='teacher')
+
+    def __init__(self, username, email, password, role='teacher'):
+        self.username = username
+        self.email = email
+        self.password = bcrypt.hashpw(password.encode('utf-8'), bcrypt.gensalt()).decode('utf-8')
+        self.role = role
